Reorganize NapCat expand APIs by category
