+++ falowp-bot-adapter-nc/src/main/kotlin/com/blr19c/falowp/bot/adapter/nc/expand/NapCatGroupApiExpand.kt
@@ -6,10 +6,12 @@ import com.blr19c.falowp.bot.adapter.nc.api.NapCatBotApi
 import com.blr19c.falowp.bot.adapter.nc.api.NapCatBotApiSupport
 import com.blr19c.falowp.bot.adapter.nc.message.NapCatMessage
 import com.blr19c.falowp.bot.system.api.ReceiveMessage
+import com.blr19c.falowp.bot.system.json.Json
 import com.fasterxml.jackson.annotation.JsonProperty
+import tools.jackson.databind.JsonNode
 
 /**
- * NapCatGroupApiExpand
+ * NapCatGroupApiExpand 群组API
  */
 class NapCatGroupApiExpand {
 
@@ -415,6 +417,47 @@ class NapCatGroupApiExpand {
         val shutUpTime: Long
     )
 
+    /**
+     * 相册信息
+     */
+    data class Album(
+        /**
+         * 相册 ID
+         */
+        @field:JsonProperty("album_id")
+        val albumId: String,
+        /**
+         * 相册名称
+         */
+        @field:JsonProperty("album_name")
+        val albumName: String,
+        /**
+         * 相册封面 URL
+         */
+        @field:JsonProperty("cover_url")
+        val coverUrl: String,
+        /**
+         * 创建时间（Unix 时间戳，秒）
+         */
+        @field:JsonProperty("create_time")
+        val createTime: Long
+    )
+
+    /**
+     * 相册媒体信息
+     */
+    data class AlbumMedia(
+        /**
+         * 媒体 ID
+         */
+        @field:JsonProperty("media_id")
+        val mediaId: String,
+        /**
+         * 媒体访问 URL
+         */
+        @field:JsonProperty("url")
+        val url: String
+    )
 
 }
 
@@ -422,6 +465,9 @@ class NapCatGroupApiExpand {
  * 删除群公告
  *
  * 删除群聊中的公告
+ *
+ * @param groupId 群组ID
+ * @param noticeId 公告ID
  */
 suspend fun NapCatBotApi.delGroupNotice(groupId: String, noticeId: String) {
     apiRequestUnit("_del_group_notice", mapOf("group_id" to groupId, "notice_id" to noticeId))
@@ -431,6 +477,8 @@ suspend fun NapCatBotApi.delGroupNotice(groupId: String, noticeId: String) {
  * 获取群公告
  *
  * 获取指定群聊中的公告列表
+ *
+ * @param groupId 群组ID
  */
 suspend fun NapCatBotApi.getGroupNotice(groupId: String): List<NapCatGroupApiExpand.GroupNoticeItemItem> {
     return apiRequest("_get_group_notice", mapOf("group_id" to groupId))
@@ -440,6 +488,11 @@ suspend fun NapCatBotApi.getGroupNotice(groupId: String): List<NapCatGroupApiExp
  * 移出精华消息
  *
  * 将一条消息从群精华消息列表中移出
+ *
+ * @param messageId 消息ID
+ * @param msgSeq 消息序列号
+ * @param msgRandom 随机值
+ * @param groupId 群组ID
  */
 suspend fun NapCatBotApi.deleteEssenceMsg(
     messageId: Long? = null,
@@ -457,6 +510,8 @@ suspend fun NapCatBotApi.deleteEssenceMsg(
  * 获取群精华消息
  *
  * 获取指定群聊中的精华消息列表
+ *
+ * @param groupId 群组ID
  */
 suspend fun NapCatBotApi.getEssenceMsgList(groupId: String): List<NapCatGroupApiExpand.EssenceMsg> {
     return apiRequest("get_essence_msg_list", mapOf("group_id" to groupId))
@@ -466,6 +521,8 @@ suspend fun NapCatBotApi.getEssenceMsgList(groupId: String): List<NapCatGroupApi
  * 获取群详细信息
  *
  * 获取群聊的详细信息，包括成员数、最大成员数等
+ *
+ * @param groupId 群组ID
  */
 suspend fun NapCatBotApi.getGroupDetailInfo(groupId: String): NapCatGroupApiExpand.GroupDetailInfo {
     return apiRequest("get_group_detail_info", mapOf("group_id" to groupId))
@@ -491,6 +548,8 @@ suspend fun NapCatBotApi.getGroupIgnoredNotifies(): NapCatGroupApiExpand.GroupIg
  * 获取群信息
  *
  * 获取群聊的基本信息
+ *
+ * @param groupId 群组ID
  */
 suspend fun NapCatBotApi.getGroupInfo(groupId: String): NapCatGroupApiExpand.GroupInfo {
     return apiRequest("get_group_info", mapOf("group_id" to groupId))
@@ -500,6 +559,8 @@ suspend fun NapCatBotApi.getGroupInfo(groupId: String): NapCatGroupApiExpand.Gro
  * 获取群列表
  *
  * 获取当前账号的群聊列表
+ *
+ * @param noCache 是否禁用缓存
  */
 suspend fun NapCatBotApi.getGroupList(noCache: Boolean? = null): List<NapCatGroupApiExpand.GroupItem> {
     return apiRequest("get_group_list", mapOf("no_cache" to noCache))
@@ -509,6 +570,10 @@ suspend fun NapCatBotApi.getGroupList(noCache: Boolean? = null): List<NapCatGrou
  * 获取群成员信息
  *
  * 获取群聊中指定成员的信息
+ *
+ * @param groupId 群组ID
+ * @param userId 用户ID
+ * @param noCache 是否禁用缓存
  */
 suspend fun NapCatBotApi.getGroupMemberInfo(
     groupId: String,
@@ -522,6 +587,9 @@ suspend fun NapCatBotApi.getGroupMemberInfo(
  * 获取群成员列表
  *
  * 获取群聊中的所有成员列表
+ *
+ * @param groupId 群组ID
+ * @param noCache 是否禁用缓存
  */
 suspend fun NapCatBotApi.getGroupMemberList(
     groupId: String,
@@ -535,6 +603,10 @@ suspend fun NapCatBotApi.getGroupMemberList(
  * 设置群待办
  *
  * 将指定消息设置为群待办
+ *
+ * @param groupId 群组ID
+ * @param messageId 消息ID
+ * @param messageSeq 消息序列号
  */
 suspend fun NapCatBotApi.setGroupTodo(groupId: String, messageId: String? = null, messageSeq: String? = null) {
     apiRequestUnit(
@@ -545,6 +617,8 @@ suspend fun NapCatBotApi.setGroupTodo(groupId: String, messageId: String? = null
 
 /**
  * 获取群禁言列表
+ *
+ * @param groupId 群组ID
  */
 suspend fun NapCatBotApi.getGroupShutList(groupId: String): List<NapCatGroupApiExpand.GroupShut> {
     return apiRequest("get_group_shut_list", mapOf("group_id" to groupId))
@@ -554,6 +628,8 @@ suspend fun NapCatBotApi.getGroupShutList(groupId: String): List<NapCatGroupApiE
  * 设置精华消息
  *
  * 将一条消息设置为群精华消息
+ *
+ * @param messageId 消息ID
  */
 suspend fun NapCatBotApi.setEssenceMsg(messageId: Long) {
     apiRequestUnit("set_essence_msg", mapOf("message_id" to messageId))
@@ -563,6 +639,11 @@ suspend fun NapCatBotApi.setEssenceMsg(messageId: Long) {
  * 处理加群请求
  *
  * 同意或拒绝加群请求或邀请
+ *
+ * @param flag 请求标识
+ * @param approve 是否同意
+ * @param reason 拒绝理由
+ * @param count 处理数量
  */
 suspend fun NapCatBotApi.setGroupAddRequest(
     flag: String,
@@ -580,6 +661,10 @@ suspend fun NapCatBotApi.setGroupAddRequest(
  * 设置群管理员
  *
  * 设置或取消群聊中的管理员
+ *
+ * @param groupId 群组ID
+ * @param userId 用户ID
+ * @param enable 是否启用
  */
 suspend fun NapCatBotApi.setGroupAdmin(groupId: String, userId: String, enable: Boolean? = null) {
     apiRequestUnit("set_group_admin", mapOf("group_id" to groupId, "user_id" to userId, "enable" to enable))
@@ -589,6 +674,10 @@ suspend fun NapCatBotApi.setGroupAdmin(groupId: String, userId: String, enable: 
  * 群组禁言
  *
  * 禁言群聊中的指定成员
+ *
+ * @param groupId 群组ID
+ * @param userId 用户ID
+ * @param duration 禁言时长
  */
 suspend fun NapCatBotApi.setGroupBan(groupId: String, userId: String, duration: Long) {
     apiRequestUnit("set_group_ban", mapOf("group_id" to groupId, "user_id" to userId, "duration" to duration))
@@ -598,6 +687,10 @@ suspend fun NapCatBotApi.setGroupBan(groupId: String, userId: String, duration: 
  * 设置群名片
  *
  * 设置群聊中指定成员的群名片
+ *
+ * @param groupId 群组ID
+ * @param userId 用户ID
+ * @param card 群名片
  */
 suspend fun NapCatBotApi.setGroupCard(groupId: String, userId: String, card: String? = null) {
     apiRequestUnit("set_group_card", mapOf("group_id" to groupId, "user_id" to userId, "card" to card))
@@ -607,6 +700,10 @@ suspend fun NapCatBotApi.setGroupCard(groupId: String, userId: String, card: Str
  * 群组踢人
  *
  * 将指定成员踢出群聊
+ *
+ * @param groupId 群组ID
+ * @param userId 用户ID
+ * @param rejectAddRequest 是否拒绝再次申请
  */
 suspend fun NapCatBotApi.setGroupKick(groupId: String, userId: String, rejectAddRequest: Boolean? = null) {
     apiRequestUnit(
@@ -619,6 +716,9 @@ suspend fun NapCatBotApi.setGroupKick(groupId: String, userId: String, rejectAdd
  * 退出群组
  *
  * 退出或解散指定群聊
+ *
+ * @param groupId 群组ID
+ * @param isDismiss 是否解散
  */
 suspend fun NapCatBotApi.setGroupLeave(groupId: String, isDismiss: Boolean? = null) {
     apiRequestUnit("set_group_leave", mapOf("group_id" to groupId, "is_dismiss" to isDismiss))
@@ -628,6 +728,9 @@ suspend fun NapCatBotApi.setGroupLeave(groupId: String, isDismiss: Boolean? = nu
  * 设置群名称
  *
  * 修改指定群聊的名称
+ *
+ * @param groupId 群组ID
+ * @param groupName 群名称
  */
 suspend fun NapCatBotApi.setGroupName(groupId: String, groupName: String) {
     apiRequestUnit("set_group_name", mapOf("group_id" to groupId, "group_name" to groupName))
@@ -637,7 +740,233 @@ suspend fun NapCatBotApi.setGroupName(groupId: String, groupName: String) {
  * 全员禁言
  *
  * 开启或关闭指定群聊的全员禁言
+ *
+ * @param groupId 群组ID
+ * @param enable 是否启用
  */
 suspend fun NapCatBotApi.setGroupWholeBan(groupId: String, enable: Boolean? = null) {
     apiRequestUnit("set_group_whole_ban", mapOf("group_id" to groupId, "enable" to enable))
+}
+
+/**
+ * 删除群相册媒体
+ *
+ * @param groupId 群组ID
+ * @param albumId 相册ID
+ * @param lloc 媒体位置标识
+ */
+suspend fun NapCatBotApi.delGroupAlbumMedia(groupId: String, albumId: String, lloc: String) {
+    apiRequestUnit("del_group_album_media", mapOf("group_id" to groupId, "album_id" to albumId, "lloc" to lloc))
+}
+
+/**
+ * 发表群相册评论
+ *
+ * @param groupId 群组ID
+ * @param albumId 相册ID
+ * @param lloc 媒体位置标识
+ * @param content 评论内容
+ */
+suspend fun NapCatBotApi.doGroupAlbumComment(groupId: String, albumId: String, lloc: String, content: String) {
+    apiRequestUnit(
+        "do_group_album_comment",
+        mapOf("group_id" to groupId, "album_id" to albumId, "lloc" to lloc, "content" to content)
+    )
+}
+
+/**
+ * 获取群相册媒体列表
+ *
+ * @param groupId 群组ID
+ * @param albumId 相册ID
+ * @param attachInfo 附加信息
+ */
+suspend fun NapCatBotApi.getGroupAlbumMediaList(
+    groupId: String,
+    albumId: String,
+    attachInfo: String
+): List<NapCatGroupApiExpand.AlbumMedia> {
+    val data = apiRequest<JsonNode>(
+        "get_group_album_media_list",
+        mapOf("group_id" to groupId, "album_id" to albumId, "attach_info" to attachInfo)
+    )
+    return Json.convertValue(data.path("media_list"))
+}
+
+/**
+ * 获取群详细信息 (扩展)
+ *
+ * @param groupId 群组ID
+ */
+suspend fun NapCatBotApi.getGroupInfoEx(groupId: String): JsonNode {
+    return apiRequest("get_group_info_ex", mapOf("group_id" to groupId))
+}
+
+/**
+ * 获取群相册列表
+ *
+ * @param groupId 群组ID
+ */
+suspend fun NapCatBotApi.getQunAlbumList(groupId: String): List<NapCatGroupApiExpand.Album> {
+    return apiRequest("get_qun_album_list", mapOf("group_id" to groupId))
+}
+
+/**
+ * 群打卡
+ *
+ * @param groupId 群组ID
+ */
+suspend fun NapCatBotApi.sendGroupSign(groupId: String) {
+    apiRequestUnit("send_group_sign", mapOf("group_id" to groupId))
+}
+
+/**
+ * 设置群加群选项
+ *
+ * @param groupId 群组ID
+ * @param addType 加群方式
+ * @param groupQuestion 验证问题
+ * @param groupAnswer 验证答案
+ */
+suspend fun NapCatBotApi.setGroupAddOption(
+    groupId: String,
+    addType: Long,
+    groupQuestion: String? = null,
+    groupAnswer: String? = null
+) {
+    apiRequestUnit(
+        "set_group_add_option",
+        mapOf(
+            "group_id" to groupId,
+            "add_type" to addType,
+            "group_question" to groupQuestion,
+            "group_answer" to groupAnswer
+        )
+    )
+}
+
+/**
+ * 点赞群相册媒体
+ *
+ * @param groupId 群组ID
+ * @param albumId 相册ID
+ * @param lloc 媒体位置标识
+ * @param id 媒体ID
+ * @param set 是否点赞
+ */
+suspend fun NapCatBotApi.setGroupAlbumMediaLike(
+    groupId: String,
+    albumId: String,
+    lloc: String,
+    id: String,
+    set: Boolean
+) {
+    apiRequestUnit(
+        "set_group_album_media_like",
+        mapOf("group_id" to groupId, "album_id" to albumId, "lloc" to lloc, "id" to id, "set" to set)
+    )
+}
+
+/**
+ * 设置群备注
+ *
+ * 设置群备注
+ *
+ * @param groupId 群组ID
+ * @param remark 群备注
+ */
+suspend fun NapCatBotApi.setGroupRemark(groupId: String, remark: String) {
+    apiRequestUnit("set_group_remark", mapOf("group_id" to groupId, "remark" to remark))
+}
+
+/**
+ * 设置群机器人加群选项
+ *
+ * @param groupId 群组ID
+ * @param robotMemberSwitch 机器人加群开关
+ * @param robotMemberExamine 机器人加群审核
+ */
+suspend fun NapCatBotApi.setGroupRobotAddOption(
+    groupId: String,
+    robotMemberSwitch: Long? = null,
+    robotMemberExamine: Long? = null
+) {
+    apiRequestUnit(
+        "set_group_robot_add_option",
+        mapOf(
+            "group_id" to groupId,
+            "robot_member_switch" to robotMemberSwitch,
+            "robot_member_examine" to robotMemberExamine
+        )
+    )
+}
+
+/**
+ * 设置群搜索选项
+ *
+ * @param groupId 群组ID
+ * @param noCodeFingerOpen 是否开启群号检索
+ * @param noFingerOpen 是否开启群名检索
+ */
+suspend fun NapCatBotApi.setGroupSearch(groupId: String, noCodeFingerOpen: Long? = null, noFingerOpen: Long? = null) {
+    apiRequestUnit(
+        "set_group_search",
+        mapOf("group_id" to groupId, "no_code_finger_open" to noCodeFingerOpen, "no_finger_open" to noFingerOpen)
+    )
+}
+
+/**
+ * 群打卡
+ *
+ * @param groupId 群组ID
+ */
+suspend fun NapCatBotApi.setGroupSign(groupId: String) {
+    apiRequestUnit("set_group_sign", mapOf("group_id" to groupId))
+}
+
+/**
+ * 上传图片到群相册
+ *
+ * @param groupId 群组ID
+ * @param albumId 相册ID
+ * @param albumName 相册名称
+ * @param file 图片文件
+ */
+suspend fun NapCatBotApi.uploadImageToQunAlbum(groupId: String, albumId: String, albumName: String, file: String) {
+    apiRequestUnit(
+        "upload_image_to_qun_album",
+        mapOf("group_id" to groupId, "album_id" to albumId, "album_name" to albumName, "file" to file)
+    )
+}
+
+/**
+ * 批量踢出群成员
+ *
+ * 从指定群聊中批量踢出多个成员
+ *
+ * @param groupId 群组ID
+ * @param userId 用户ID列表
+ * @param rejectAddRequest 是否拒绝再次申请
+ */
+suspend fun NapCatBotApi.setGroupKickMembers(groupId: String, userId: List<String>, rejectAddRequest: Boolean? = null) {
+    apiRequestUnit(
+        "set_group_kick_members",
+        mapOf("group_id" to groupId, "user_id" to userId, "reject_add_request" to rejectAddRequest)
+    )
+}
+
+/**
+ * 设置专属头衔
+ *
+ * 设置群聊中指定成员的专属头衔
+ *
+ * @param groupId 群组ID
+ * @param userId 用户ID
+ * @param specialTitle 专属头衔
+ */
+suspend fun NapCatBotApi.setGroupSpecialTitle(groupId: String, userId: String, specialTitle: String) {
+    apiRequestUnit(
+        "set_group_special_title",
+        mapOf("group_id" to groupId, "user_id" to userId, "special_title" to specialTitle)
+    )
 }
