+++ falowp-bot-adapter-nc/src/main/kotlin/com/blr19c/falowp/bot/adapter/nc/expand/NapCatGuildApiExpand.kt
@@ -6,7 +6,7 @@ import com.blr19c.falowp.bot.adapter.nc.api.NapCatBotApi
 import tools.jackson.databind.JsonNode
 
 /**
- * NapCatGuildApiExpand
+ * NapCatGuildApiExpand 频道API
  */
 class NapCatGuildApiExpand {
 }
