+++ falowp-bot-adapter-nc/src/main/kotlin/com/blr19c/falowp/bot/adapter/nc/expand/NapCatSystemApiExpand.kt
@@ -7,7 +7,7 @@ import com.fasterxml.jackson.annotation.JsonProperty
 import tools.jackson.databind.JsonNode
 
 /**
- * NapCatSystemApiExpand
+ * NapCatSystemApiExpand 系统API
  */
 class NapCatSystemApiExpand {
     /**
@@ -35,6 +35,17 @@ class NapCatSystemApiExpand {
          */
         @field:JsonProperty("token")
         val token: Long
+    )
+
+    /**
+     * ClientKey
+     */
+    data class ClientKey(
+        /**
+         * 客户端Key
+         */
+        @field:JsonProperty("clientkey")
+        val clientKey: String?
     )
 
     /**
@@ -299,6 +310,85 @@ class NapCatSystemApiExpand {
         val appVersion: String
     )
 
+    /**
+     * MiniAppArk
+     */
+    data class MiniAppArk(
+        /**
+         * data
+         */
+        @field:JsonProperty("data")
+        val data: String
+    )
+
+    /**
+     * RkeyItemItem
+     */
+    data class RKeyItemItem(
+        /**
+         * 类型 (private/group)
+         */
+        @field:JsonProperty("type")
+        val type: String,
+        /**
+         * RKey
+         */
+        @field:JsonProperty("rkey")
+        val rKey: String,
+        /**
+         * 创建时间
+         */
+        @field:JsonProperty("created_at")
+        val createdAt: Long,
+        /**
+         * 有效期
+         */
+        @field:JsonProperty("ttl")
+        val ttl: Long
+    )
+
+    /**
+     * RkeyServer
+     */
+    data class RkeyServer(
+        /**
+         * 私聊 RKey
+         */
+        @field:JsonProperty("private_rkey")
+        val privateRKey: String?,
+        /**
+         * 群聊 RKey
+         */
+        @field:JsonProperty("group_rkey")
+        val groupRKey: String?,
+        /**
+         * 过期时间
+         */
+        @field:JsonProperty("expired_time")
+        val expiredTime: Long?,
+        /**
+         * 名称
+         */
+        @field:JsonProperty("name")
+        val name: String
+    )
+
+    /**
+     * NcGetUserStatus
+     */
+    data class NcGetUserStatus(
+        /**
+         * 在线状态
+         */
+        @field:JsonProperty("status")
+        val status: Long,
+        /**
+         * 扩展状态
+         */
+        @field:JsonProperty("ext_status")
+        val extStatus: Long
+    )
+
 }
 
 /**
@@ -332,6 +422,8 @@ suspend fun NapCatBotApi.cleanCache() {
  * 获取登录凭证
  *
  * 获取登录凭证
+ *
+ * @param domain 域名
  */
 suspend fun NapCatBotApi.getCredentials(domain: String): NapCatSystemApiExpand.Credentials {
     return apiRequest("get_credentials", mapOf("domain" to domain))
@@ -350,6 +442,8 @@ suspend fun NapCatBotApi.getCsrfToken(): NapCatSystemApiExpand.CsrfToken {
  * 获取可疑好友申请
  *
  * 获取系统的可疑好友申请列表
+ *
+ * @param count 获取数量
  */
 suspend fun NapCatBotApi.getDoubtFriendsAddRequest(count: Long): List<NapCatSystemApiExpand.DoubtFriends> {
     return apiRequest("get_doubt_friends_add_request", mapOf("count" to count))
@@ -359,6 +453,8 @@ suspend fun NapCatBotApi.getDoubtFriendsAddRequest(count: Long): List<NapCatSyst
  * 获取群系统消息
  *
  * 获取群系统消息
+ *
+ * @param count 获取数量
  */
 suspend fun NapCatBotApi.getGroupSystemMsg(count: Long): NapCatSystemApiExpand.GroupSystemMsg {
     return apiRequest("get_group_system_msg", mapOf("count" to count))
@@ -404,6 +500,9 @@ suspend fun NapCatBotApi.ncGetPacketStatus(): JsonNode {
  * 处理可疑好友申请
  *
  * 同意或拒绝系统的可疑好友申请
+ *
+ * @param flag 请求标识
+ * @param approve 是否同意
  */
 suspend fun NapCatBotApi.setDoubtFriendsAddRequest(flag: String, approve: Boolean) {
     apiRequestUnit("set_doubt_friends_add_request", mapOf("flag" to flag, "approve" to approve))
@@ -416,4 +515,136 @@ suspend fun NapCatBotApi.setDoubtFriendsAddRequest(flag: String, approve: Boolea
  */
 suspend fun NapCatBotApi.setRestart() {
     apiRequestUnit("set_restart")
+}
+
+/**
+ * 退出登录
+ */
+suspend fun NapCatBotApi.botExit() {
+    apiRequestUnit("bot_exit")
+}
+
+/**
+ * 获取自定义表情
+ *
+ * @param count 获取数量
+ */
+suspend fun NapCatBotApi.fetchCustomFace(count: Long): List<String> {
+    return apiRequest("fetch_custom_face", mapOf("count" to count))
+}
+
+/**
+ * 获取收藏列表
+ *
+ * @param category 收藏分类
+ * @param count 获取数量
+ */
+suspend fun NapCatBotApi.getCollectionList(category: String, count: String): String {
+    return apiRequest("get_collection_list", mapOf("category" to category, "count" to count))
+}
+
+/**
+ * 获取小程序 Ark
+ */
+suspend fun NapCatBotApi.getMiniAppArk(): NapCatSystemApiExpand.MiniAppArk {
+    return apiRequest("get_mini_app_ark")
+}
+
+/**
+ * 获取扩展 RKey
+ */
+suspend fun NapCatBotApi.getRKey(): List<NapCatSystemApiExpand.RKeyItemItem> {
+    return apiRequest("get_rkey")
+}
+
+/**
+ * 获取 RKey 服务器
+ */
+suspend fun NapCatBotApi.getRKeyServer(): NapCatSystemApiExpand.RkeyServer {
+    return apiRequest("get_rkey_server")
+}
+
+/**
+ * 获取机器人 UIN 范围
+ */
+suspend fun NapCatBotApi.getRobotUinRange(): List<String> {
+    return apiRequest("get_robot_uin_range")
+}
+
+/**
+ * 获取 RKey
+ */
+suspend fun NapCatBotApi.ncGetRKey(): List<String> {
+    return apiRequest("nc_get_rkey")
+}
+
+/**
+ * 获取用户在线状态
+ *
+ * @param userId 用户ID
+ */
+suspend fun NapCatBotApi.ncGetUserStatus(userId: String): NapCatSystemApiExpand.NcGetUserStatus {
+    return apiRequest("nc_get_user_status", mapOf("user_id" to userId))
+}
+
+/**
+ * 发送原始数据包
+ *
+ * @param cmd 命令字
+ * @param data 数据内容
+ * @param rsp 响应配置
+ */
+suspend fun NapCatBotApi.sendPacket(cmd: String, data: String, rsp: String) {
+    apiRequestUnit("send_packet", mapOf("cmd" to cmd, "data" to data, "rsp" to rsp))
+}
+
+/**
+ * 设置输入状态
+ *
+ * @param userId 用户ID
+ * @param eventType 事件类型
+ */
+suspend fun NapCatBotApi.setInputStatus(userId: String, eventType: Long) {
+    apiRequestUnit("set_input_status", mapOf("user_id" to userId, "event_type" to eventType))
+}
+
+/**
+ * 创建收藏
+ *
+ * @param rawData 原始数据
+ * @param brief 简要描述
+ */
+suspend fun NapCatBotApi.createCollection(rawData: String, brief: String) {
+    apiRequestUnit("create_collection", mapOf("rawData" to rawData, "brief" to brief))
+}
+
+/**
+ * 获取ClientKey
+ *
+ * 获取当前登录帐号的ClientKey
+ */
+suspend fun NapCatBotApi.getClientKey(): NapCatSystemApiExpand.ClientKey {
+    return apiRequest("get_clientkey")
+}
+
+/**
+ * 图片 OCR 识别
+ *
+ * 识别图片中的文字内容(仅Windows端支持)
+ *
+ * @param image 图片数据
+ */
+suspend fun NapCatBotApi.ocrImage(image: String) {
+    apiRequestUnit("ocr_image", mapOf("image" to image))
+}
+
+/**
+ * 英文单词翻译
+ *
+ * 将英文单词列表翻译为中文
+ *
+ * @param words 英文单词列表
+ */
+suspend fun NapCatBotApi.translateEn2zh(words: List<String>) {
+    apiRequestUnit("translate_en2zh", mapOf("words" to words))
 }
